refactor spel expression suggester
