+++ utils/utils/src/main/scala/pl/touk/nussknacker/engine/util/CaretPosition2d.scala
@@ -1,0 +1,12 @@
+package pl.touk.nussknacker.engine.util
+
+import io.circe.generic.JsonCodec
+
+@JsonCodec
+final case class CaretPosition2d(row: Int, column: Int) {
+
+  def normalizedCaretPosition(inputValue: String): Int = {
+    inputValue.split("\n").take(row).map(_.length).sum + row + column
+  }
+
+}
